Add router to separate pages
식단 페이지를 별도로 구성하기 위해 라우터를 적용했습니다.
