+++ .eslintrc.js
@@ -49,7 +49,7 @@ module.exports = {
     'key-spacing': ['error', { mode: 'strict' }],
     'jsx-a11y/label-has-associated-control': ['error', { assert: 'either' }],
     'linebreak-style': 'off',
-    'import/no-extraneous-dependencies': ['error', { devDependencies: true, optionalDependencies: false, peerDependencies: false }],
+    'import/no-extraneous-dependencies': 'off',
     'react/prop-types': 'off',
     'react/react-in-jsx-scope': 'off',
   },
