+++ src/DiaryPage.jsx
@@ -1,0 +1,7 @@
+export default function DiaryPage() {
+  return (
+    <div>
+      <h1>식단 페이지</h1>
+    </div>
+  );
+}
