+++ src/index.jsx
@@ -2,6 +2,8 @@ import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
 
+import { BrowserRouter } from 'react-router-dom';
+
 import App from './App';
 
 import store from './store';
@@ -9,7 +11,9 @@ import store from './store';
 ReactDOM.render(
   (
     <Provider store={store}>
-      <App />
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
     </Provider>
   ),
   document.getElementById('app'),
